requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-heroku==0.3.1
django-phone-field==1.8.0
djangorestframework==3.11.0
export==0.1.2
gunicorn==20.0.4
mysqlclient==1.4.6
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
